Create a common module for SGE related commands
